@@ -1,19 +1,13 @@
 module ActiveModel
   class Serializer
     module Adapter
-      class Json < Base
-        extend ActiveSupport::Autoload
-        autoload :FragmentCache
-
-        def serializable_hash(options = nil)
-          options ||= {}
-          { root => Attributes.new(serializer, instance_options).serializable_hash(options) }
+      class Json < DelegateClass(ActiveModelSerializers::Adapter::Json)
+        def initialize(serializer, options = {})
+          super(ActiveModelSerializers::Adapter::Json.new(serializer, options))
         end
-
-        private
-
-        def fragment_cache(cached_hash, non_cached_hash)
-          ActiveModel::Serializer::Adapter::Json::FragmentCache.new.fragment_cache(cached_hash, non_cached_hash)
+        class << self
+          extend ActiveModelSerializers::Deprecate
+          deprecate :new, 'ActiveModelSerializers::Adapter::Json.new'
         end
       end
     end
