@@ -1,9 +1,10 @@
 module ActiveModel
   class Serializer
     # @api private
-    class BelongsToReflection < SingularReflection
-      def macro
-        :belongs_to
+    class BelongsToReflection < Reflection
+      # @api private
+      def foreign_key_on
+        :self
       end
     end
   end
