@@ -1,9 +1,9 @@
 module ActiveModel
   class Serializer
     # @api private
-    class HasManyReflection < CollectionReflection
-      def macro
-        :has_many
+    class HasManyReflection < Reflection
+      def collection?
+        true
       end
     end
   end
