@@ -1,5 +1,5 @@
 module ActiveModel
   class Serializer
-    VERSION = '0.10.0.rc3'
+    VERSION = '0.10.8'.freeze
   end
 end
