@@ -1,59 +1,61 @@
-require 'logger'
 require 'active_model'
 require 'active_support'
-require 'action_controller'
-require 'action_controller/railtie'
+require 'active_support/core_ext/object/with_options'
+require 'active_support/core_ext/string/inflections'
+require 'active_support/json'
 module ActiveModelSerializers
-  mattr_accessor :logger
-  self.logger = Rails.logger || Logger.new(IO::NULL)
-
   extend ActiveSupport::Autoload
-  autoload :Model
+  eager_autoload do
+    autoload :Model
+    autoload :Callbacks
+    autoload :SerializableResource
+    autoload :SerializationContext
+    autoload :Logging
+    autoload :Test
+    autoload :Adapter
+    autoload :JsonPointer
+    autoload :Deprecate
+    autoload :LookupChain
+    autoload :Deserialization
+  end
 
-  module_function
+  class << self; attr_accessor :logger; end
+  self.logger = ActiveSupport::TaggedLogging.new(ActiveSupport::Logger.new(STDOUT))
 
-  # @note
-  #   ```ruby
-  #   private
-  #
-  #   attr_reader :resource, :adapter_opts, :serializer_opts
-  #   ```
-  #
-  #   Will generate a warning, though it shouldn't.
-  #   There's a bug in Ruby for this: https://bugs.ruby-lang.org/issues/10967
-  #
-  #   We can use +ActiveModelSerializers.silence_warnings+ as a
-  #   'safety valve' for unfixable or not-worth-fixing warnings,
-  #   and keep our app warning-free.
-  #
-  #   ```ruby
-  #   private
-  #
-  #   ActiveModelSerializers.silence_warnings do
-  #     attr_reader :resource, :adapter_opts, :serializer_opts
-  #   end
-  #   ```
-  #
-  #   or, as specific stopgap, define the attrs in the protected scope.
-  def silence_warnings
-    verbose = $VERBOSE
+  def self.config
+    ActiveModel::Serializer.config
+  end
+
+  # The file name and line number of the caller of the caller of this method.
+  def self.location_of_caller
+    caller[1] =~ /(.*?):(\d+).*?$/i
+    file = Regexp.last_match(1)
+    lineno = Regexp.last_match(2).to_i
+
+    [file, lineno]
+  end
+
+  # Memoized default include directive
+  # @return [JSONAPI::IncludeDirective]
+  def self.default_include_directive
+    @default_include_directive ||= JSONAPI::IncludeDirective.new(config.default_includes, allow_wildcard: true)
+  end
+
+  def self.silence_warnings
+    original_verbose = $VERBOSE
     $VERBOSE = nil
     yield
   ensure
-    $VERBOSE = verbose
+    $VERBOSE = original_verbose
   end
-end
 
-require 'active_model/serializer'
-require 'active_model/serializable_resource'
-require 'active_model/serializer/version'
-
-require 'action_controller/serialization'
-ActiveSupport.on_load(:action_controller) do
-  include ::ActionController::Serialization
-  ActionDispatch::Reloader.to_prepare do
-    ActiveModel::Serializer.serializers_cache.clear
+  def self.eager_load!
+    super
+    ActiveModel::Serializer.eager_load!
   end
-end
 
-require 'active_model/serializer/railtie'
+  require 'active_model/serializer/version'
+  require 'active_model/serializer'
+  require 'active_model/serializable_resource'
+  require 'active_model_serializers/railtie' if defined?(::Rails::Railtie)
+end
