@@ -1,63 +1,92 @@
-verbose = $VERBOSE
-$VERBOSE = nil
 class Model < ActiveModelSerializers::Model
-  FILE_DIGEST = Digest::MD5.hexdigest(File.open(__FILE__).read)
+  rand(2).zero? && derive_attributes_from_names_and_fix_accessors
 
-  ### Helper methods, not required to be serializable
+  attr_writer :id
 
-  # Convenience when not adding @attributes readers and writers
-  def method_missing(meth, *args)
-    if meth.to_s =~ /^(.*)=$/
-      attributes[$1.to_sym] = args[0]
-    elsif attributes.key?(meth)
-      attributes[meth]
-    else
-      super
+  # At this time, just for organization of intent
+  class_attribute :association_names
+  self.association_names = []
+
+  def self.associations(*names)
+    self.association_names |= names.map(&:to_sym)
+    # Silence redefinition of methods warnings
+    ActiveModelSerializers.silence_warnings do
+      attr_accessor(*names)
     end
   end
 
-  def cache_key_with_digest
-    "#{cache_key}/#{FILE_DIGEST}"
+  def associations
+    association_names.each_with_object({}) do |association_name, result|
+      result[association_name] = public_send(association_name).freeze
+    end.with_indifferent_access.freeze
   end
+
+  def attributes
+    super.except(*association_names)
+  end
+end
+
+# see
+# https://github.com/rails/rails/blob/4-2-stable/activemodel/lib/active_model/errors.rb
+# The below allows you to do:
+#
+#   model = ModelWithErrors.new
+#   model.validate!            # => ["cannot be nil"]
+#   model.errors.full_messages # => ["name cannot be nil"]
+class ModelWithErrors < Model
+  attributes :name
 end
 
 class Profile < Model
+  attributes :name, :description
+  associations :comments
 end
-
 class ProfileSerializer < ActiveModel::Serializer
   attributes :name, :description
-
-  def arguments_passed_in?
-    instance_options[:my_options] == :accessible
-  end
 end
-
 class ProfilePreviewSerializer < ActiveModel::Serializer
   attributes :name
 end
 
-Post     = Class.new(Model)
-Like     = Class.new(Model)
-Author   = Class.new(Model)
-Bio      = Class.new(Model)
-Blog     = Class.new(Model)
-Role     = Class.new(Model)
-User     = Class.new(Model)
-Location = Class.new(Model)
-Place    = Class.new(Model)
-Tag      = Class.new(Model)
-VirtualValue = Class.new(Model)
-Comment = Class.new(Model) do
-  # Uses a custom non-time-based cache key
-  def cache_key
-    "#{self.class.name.downcase}/#{self.id}"
-  end
+class Author < Model
+  attributes :name
+  associations :posts, :bio, :roles, :comments
+end
+class AuthorSerializer < ActiveModel::Serializer
+  cache key: 'writer', skip_digest: true
+  attribute :id
+  attribute :name
+
+  has_many :posts
+  has_many :roles
+  has_one :bio
+end
+class AuthorPreviewSerializer < ActiveModel::Serializer
+  attributes :id
+  has_many :posts
 end
 
-module Spam; end
-Spam::UnrelatedLink = Class.new(Model)
+class Comment < Model
+  attributes :body, :date
+  associations :post, :author, :likes
+end
+class CommentSerializer < ActiveModel::Serializer
+  cache expires_in: 1.day, skip_digest: true
+  attributes :id, :body
+  belongs_to :post
+  belongs_to :author
+end
+class CommentPreviewSerializer < ActiveModel::Serializer
+  attributes :id
 
-PostSerializer = Class.new(ActiveModel::Serializer) do
+  belongs_to :post
+end
+
+class Post < Model
+  attributes :title, :body
+  associations :author, :comments, :blog, :tags, :related
+end
+class PostSerializer < ActiveModel::Serializer
   cache key: 'post', expires_in: 0.1, skip_digest: true
   attributes :id, :title, :body
 
@@ -68,152 +97,119 @@ PostSerializer = Class.new(ActiveModel::Serializer) do
   def blog
     Blog.new(id: 999, name: 'Custom blog')
   end
-
-  def custom_options
-    instance_options
-  end
 end
-
-SpammyPostSerializer = Class.new(ActiveModel::Serializer) do
+class SpammyPostSerializer < ActiveModel::Serializer
   attributes :id
   has_many :related
+end
+class PostPreviewSerializer < ActiveModel::Serializer
+  attributes :title, :body, :id
 
-  def self.root_name
-    'posts'
-  end
+  has_many :comments, serializer: ::CommentPreviewSerializer
+  belongs_to :author, serializer: ::AuthorPreviewSerializer
+end
+class PostWithCustomKeysSerializer < ActiveModel::Serializer
+  attributes :id
+  has_many :comments, key: :reviews
+  belongs_to :author, key: :writer
+  has_one :blog, key: :site
 end
 
-CommentSerializer = Class.new(ActiveModel::Serializer) do
-  cache expires_in: 1.day, skip_digest: true
-  attributes :id, :body
-
-  belongs_to :post
-  belongs_to :author
-
-  def custom_options
-    instance_options
-  end
+class Bio < Model
+  attributes :content, :rating
+  associations :author
 end
-
-AuthorSerializer = Class.new(ActiveModel::Serializer) do
-  cache key: 'writer', skip_digest: true
-  attribute :id
-  attribute :name
-
-  has_many :posts
-  has_many :roles
-  has_one :bio
-end
-
-RoleSerializer = Class.new(ActiveModel::Serializer) do
-  cache only: [:name], skip_digest: true
-  attributes :id, :name, :description, :slug
-
-  def slug
-    "#{name}-#{id}"
-  end
-
-  belongs_to :author
-end
-
-LikeSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id, :time
-
-  belongs_to :likeable
-end
-
-LocationSerializer = Class.new(ActiveModel::Serializer) do
-  cache only: [:place], skip_digest: true
-  attributes :id, :lat, :lng
-
-  belongs_to :place
-
-  def place
-    'Nowhere'
-  end
-end
-
-PlaceSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id, :name
-
-  has_many :locations
-end
-
-BioSerializer = Class.new(ActiveModel::Serializer) do
+class BioSerializer < ActiveModel::Serializer
   cache except: [:content], skip_digest: true
   attributes :id, :content, :rating
 
   belongs_to :author
 end
 
-BlogSerializer = Class.new(ActiveModel::Serializer) do
+class Blog < Model
+  attributes :name, :type, :special_attribute
+  associations :writer, :articles
+end
+class BlogSerializer < ActiveModel::Serializer
   cache key: 'blog'
   attributes :id, :name
 
   belongs_to :writer
   has_many :articles
 end
-
-PaginatedSerializer = Class.new(ActiveModel::Serializer::CollectionSerializer) do
-  def json_key
-    'paginated'
-  end
-end
-
-AlternateBlogSerializer = Class.new(ActiveModel::Serializer) do
+class AlternateBlogSerializer < ActiveModel::Serializer
   attribute :id
   attribute :name, key: :title
 end
-
-CustomBlogSerializer = Class.new(ActiveModel::Serializer) do
+class CustomBlogSerializer < ActiveModel::Serializer
   attribute :id
   attribute :special_attribute
-
   has_many :articles
 end
 
-CommentPreviewSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id
-
-  belongs_to :post
+class Role < Model
+  attributes :name, :description, :special_attribute
+  associations :author
 end
+class RoleSerializer < ActiveModel::Serializer
+  cache only: [:name, :slug], skip_digest: true
+  attributes :id, :name, :description
+  attribute :friendly_id, key: :slug
+  belongs_to :author
 
-AuthorPreviewSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id
-
-  has_many :posts
-end
-
-PostPreviewSerializer = Class.new(ActiveModel::Serializer) do
-  def self.root_name
-    'posts'
+  def friendly_id
+    "#{object.name}-#{object.id}"
   end
-
-  attributes :title, :body, :id
-
-  has_many :comments, serializer: CommentPreviewSerializer
-  belongs_to :author, serializer: AuthorPreviewSerializer
 end
 
-PostWithTagsSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id
+class Location < Model
+  attributes :lat, :lng
+  associations :place
+end
+class LocationSerializer < ActiveModel::Serializer
+  cache only: [:address], skip_digest: true
+  attributes :id, :lat, :lng
 
-  has_many :tags
+  belongs_to :place, key: :address
+
+  def place
+    'Nowhere'
+  end
 end
 
-PostWithCustomKeysSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id
-
-  has_many :comments, key: :reviews
-  belongs_to :author, key: :writer
-  has_one :blog, key: :site
+class Place < Model
+  attributes :name
+  associations :locations
+end
+class PlaceSerializer < ActiveModel::Serializer
+  attributes :id, :name
+  has_many :locations
 end
 
-VirtualValueSerializer = Class.new(ActiveModel::Serializer) do
-  attributes :id
+class Like < Model
+  attributes :time
+  associations :likeable
+end
+class LikeSerializer < ActiveModel::Serializer
+  attributes :id, :time
+  belongs_to :likeable
+end
 
-  has_many :reviews, virtual_value: [{ id: 1 }, { id: 2 }]
-  has_one :maker, virtual_value: { id: 1 }
+module Spam
+  class UnrelatedLink < Model
+  end
+  class UnrelatedLinkSerializer < ActiveModel::Serializer
+    cache only: [:id]
+    attributes :id
+  end
+end
+
+class VirtualValue < Model; end
+class VirtualValueSerializer < ActiveModel::Serializer
+  attributes :id
+  has_many :reviews, virtual_value: [{ type: 'reviews', id: '1' },
+                                     { type: 'reviews', id: '2' }]
+  has_one :maker, virtual_value: { type: 'makers', id: '1' }
 
   def reviews
   end
@@ -222,8 +218,8 @@ VirtualValueSerializer = Class.new(ActiveModel::Serializer) do
   end
 end
 
-Spam::UnrelatedLinkSerializer = Class.new(ActiveModel::Serializer) do
-  cache only: [:id]
-  attributes :id
+class PaginatedSerializer < ActiveModel::Serializer::CollectionSerializer
+  def json_key
+    'paginated'
+  end
 end
-$VERBOSE = verbose
