@@ -2,13 +2,13 @@ require 'test_helper'
 
 module ActiveModel
   class Serializer
-    class RootTest < Minitest::Test
+    class RootTest < ActiveSupport::TestCase
       def setup
         @virtual_value = VirtualValue.new(id: 1)
       end
 
       def test_overwrite_root
-        serializer = VirtualValueSerializer.new(@virtual_value, { root: 'smth' })
+        serializer = VirtualValueSerializer.new(@virtual_value, root: 'smth')
         assert_equal('smth', serializer.json_key)
       end
 
