@@ -1,64 +1,72 @@
+# Configure Rails Environment
+ENV['RAILS_ENV'] = 'test'
 require 'bundler/setup'
 
 begin
   require 'simplecov'
-  # HACK: till https://github.com/colszowka/simplecov/pull/400 is merged and released.
-  # Otherwise you may get:
-  # simplecov-0.10.0/lib/simplecov/defaults.rb:50: warning: global variable `$ERROR_INFO' not initialized
-  require 'support/simplecov'
   AppCoverage.start
 rescue LoadError
   STDERR.puts 'Running without SimpleCov'
 end
 
+require 'pry'
 require 'timecop'
 require 'rails'
 require 'action_controller'
 require 'action_controller/test_case'
 require 'action_controller/railtie'
-require 'active_support/json'
+require 'active_model_serializers'
+# For now, we only restrict the options to serializable_hash/as_json/to_json
+# in tests, to ensure developers don't add any unsupported options.
+# There's no known benefit, at this time, to having the filtering run in
+# production when the excluded options would simply not be used.
+#
+# However, for documentation purposes, the constant
+# ActiveModel::Serializer::SERIALIZABLE_HASH_VALID_KEYS is defined
+# in the Serializer.
+ActiveModelSerializers::Adapter::Base.class_eval do
+  alias_method :original_serialization_options, :serialization_options
+
+  def serialization_options(options)
+    original_serialization_options(options)
+      .slice(*ActiveModel::Serializer::SERIALIZABLE_HASH_VALID_KEYS)
+  end
+end
 require 'fileutils'
 FileUtils.mkdir_p(File.expand_path('../../tmp/cache', __FILE__))
 
 gem 'minitest'
+require 'minitest'
 require 'minitest/autorun'
-require 'minitest/reporters'
-Minitest::Reporters.use!
-if defined?(Minitest::Test)
-  # Minitest 5
-  # https://github.com/seattlerb/minitest/blob/e21fdda9d/lib/minitest/autorun.rb
-  # https://github.com/seattlerb/minitest/blob/e21fdda9d/lib/minitest.rb#L45-L59
-else
-  # Minitest 4
-  # https://github.com/seattlerb/minitest/blob/644a52fd0/lib/minitest/autorun.rb
-  # https://github.com/seattlerb/minitest/blob/644a52fd0/lib/minitest/unit.rb#L768-L787
-  # Ensure backward compatibility with Minitest 4
-  Minitest = MiniTest unless defined?(Minitest)
-  Minitest::Test = MiniTest::Unit::TestCase
-  def Minitest.after_run(&block)
-    MiniTest::Unit.after_tests(&block)
+Minitest.backtrace_filter = Minitest::BacktraceFilter.new
+
+module TestHelper
+  module_function
+
+  def silence_warnings
+    original_verbose = $VERBOSE
+    $VERBOSE = nil
+    yield
+  ensure
+    $VERBOSE = original_verbose
   end
 end
 
-# If there's no failure info, try disabling capturing stderr:
-# `env CAPTURE_STDERR=false rake`
-# This is way easier than writing a Minitest plugin
-# for 4.x and 5.x.
-if ENV['CAPTURE_STDERR'] !~ /false|1/i
-  require 'capture_warnings'
-  CaptureWarnings.new(_fail_build = true).execute!
-end
-
-require 'active_model_serializers'
-
-require 'support/stream_capture'
-
 require 'support/rails_app'
 
-require 'support/test_case'
+# require "rails/test_help"
 
 require 'support/serialization_testing'
+
+require 'support/rails5_shims'
+
+require 'support/match_array'
 
 require 'fixtures/active_record'
 
 require 'fixtures/poro'
+
+ActiveSupport.on_load(:action_controller) do
+  $action_controller_logger = ActiveModelSerializers.logger
+  ActiveModelSerializers.logger = Logger.new(IO::NULL)
+end
